test: cover case of min/maxValues differing in size
- add test to cover the one line of index.js that was being missed
  https://codecov.io/gh/mljs/levenberg-marquardt/src/9c99709b23fa5d027fc8da0e3d228e6934d56217/src/index.js#L57
- remove unnecessary (and skipped) sinusoidal test for errorCalculation
- replace overly complicated sumOfLorentzians function with sinFunction in invalid arguments exception handling test
- move "Handling of ill-behaved functions" tests from curve.js to test.js
- rename test.js -> exceptions.js now that all of its tests are for handling of exceptional cases
  (tests related to fitting specific functions / data moved to curve.js)
