+++ src/__tests__/curve.js
@@ -130,28 +130,3 @@ describe('"Real-world" problems (noisy data)', () => {
     });
   });
 });
-
-
-describe('Handling of ill-behaved functions', () => {
-  it('Should stop and return parameterError=NaN if function evaluates to NaN after starting', () => {
-    const fourParamEq = ([a, b, c, d]) => ((t) => a + (b - a) / (1 + Math.pow(c, d) * Math.pow(t, -d)));
-    const data = {
-      x: [9.22e-12, 5.53e-11, 3.32e-10, 1.99e-9, 1.19e-8, 7.17e-8, 4.3e-7, 0.00000258, 0.0000155, 0.0000929],
-      y: [7.807, -3.74, 21.119, 2.382, 4.269, 41.57, 73.401, 98.535, 97.059, 92.147]
-    };
-    const options = {
-      damping: 0.01,
-      maxIterations: 200,
-      initialValues: [0, 100, 1, 0.1]
-      // Note: This test is identical to the other fourParamEq test, except for the
-      // damping parameter. The increased damping option leads to a case where
-      // c < 0 && d is not an integer so Math.pow(c, d) is NaN
-    };
-
-    expect(levenbergMarquardt(data, fourParamEq, options)).toBeDeepCloseTo({
-      iterations: 0,
-      parameterError: NaN,
-      parameterValues: [-64.298, 117.4022, -47.0851, -0.06148]
-    }, 3);
-  });
-});
